[TS migration] Migrate 'ReportTestUtils.js', 'deepReplaceKeysAndValuesTest.js', 'DistanceRequestUtilsTest.js', 'waitForBatchedUpdatesWithAct.js' and 'ConvertToLTRForComposerTest.js' test
