+++ tests/unit/ConvertToLTRForComposerTest.ts
