+++ tests/unit/DistanceRequestUtilsTest.ts
@@ -1,11 +1,12 @@
-import DistanceRequestUtils from '../../src/libs/DistanceRequestUtils';
+import DistanceRequestUtils from '@libs/DistanceRequestUtils';
+import type {Unit} from '@src/types/onyx/Policy';
 
 describe('DistanceRequestUtils', () => {
     describe('getDistanceRequestAmount', () => {
         test.each([
             [350, 8605.146, 'mi', 65.5],
             [561, 8605.146, 'km', 65.1],
-        ])('Correctly calculates amount %s for %s%s at a rate of %s per unit', (expectedResult, distance, unit, rate) => {
+        ] as const)('Correctly calculates amount %s for %s%s at a rate of %s per unit', (expectedResult: number, distance: number, unit: Unit, rate: number) => {
             expect(DistanceRequestUtils.getDistanceRequestAmount(distance, unit, rate)).toBe(expectedResult);
         });
     });
