+++ tests/utils/waitForBatchedUpdatesWithAct.ts
@@ -21,13 +21,11 @@ import waitForBatchedUpdates from './waitForBatchedUpdates';
  *     - You're not rendering any react components at all in your tests, but have some async logic you need to wait for e.g. Onyx.merge(). Use waitForBatchedUpdates().
  *     - You're writing UI tests but don't see any errors or warnings related to using act(). You probably don't need this in that case and should use waitForBatchedUpdates().
  *     - You're writing UI test and do see a warning about using act(), but there's no asynchronous code that needs to run inside act().
- *
- * @returns {Promise}
  */
-// eslint-disable-next-line @lwc/lwc/no-async-await
-export default async function waitForBatchedUpdatesWithAct() {
-    // eslint-disable-next-line @lwc/lwc/no-async-await
-    await act(async () => {
+async function waitForBatchedUpdatesWithAct(): Promise<void> {
+    await act(async (): Promise<void> => {
         await waitForBatchedUpdates();
     });
 }
+
+export default waitForBatchedUpdatesWithAct;
